Use `with open` or `with codecs.open` everywhere
Except when `os.devnull` is opened for forking, etc.

Ensures that all handles will be closed regardless of whether each
possible branch has a manual close statement. This makes it much more
convenient.
The downside is that, due to separate error handling, indent level has
to be increased at some points.

Check with the following regex:

    (?<!with )(?<!with codecs\.)(?<!P|_)open\(